# 01_API/src/Api/Api.yaml
openapi: 3.0.0
info:
  version: 1.0.1
  title: KPMG Pulse Mandate API
  description: Mandate API of the KPMG Pulse Platform
servers:
- url: "/api"
- url: "/mandate/api"
...
security:
- JWT: []
paths:
  "/mandate":
    get:
      tags:
      - Mandate
      summary: Get a list of filtered mandates
      description: Get a list of filtered mandates
      operationId: GetCollectionsAsync
      parameters:
      - in: query
        name: skip
        description: skip
        schema:
          type: integer
        required: false
        example: 0
      - in: query
        name: limit
        description: limit
        schema:
          type: integer
        required: false
        example: 20
      - in: query
        name: searchTerm
        description: searchTerm
        schema:
          type: string
        required: false
        example: '1000265308'
      - in: query
        name: creationDateStart
        description: CreationDateStart
        schema:
          type: string
        required: false
        example: '2023-09-01'
      - in: query
        name: creationDateEnd
        description: CreationDateEnd
        schema:
          type: string
        required: false
        example: '2023-09-01'
      - in: query
        name: modificationDateStart
        description: modificationDateStart
        schema:
          type: string
        required: false
        example: '2023-09-01'
      - in: query
        name: modificationDateEnd
        description: modificationDateEnd
        schema:
          type: string
        required: false
        example: '2023-09-01'
      - in: query
        name: statusCodes
        description: StatusCodes
        schema:
          type: array
          items:
            type: integer
        explode: true
        required: false
      - in: query
        name: sortCriteria
        description: sortCriteria
        schema:
          type: string
          enum:
          - ErpId
          - Name
          - AccountNumber
          - BankName
          - CreationDate
          - ModificationDate
          - Status
        required: false
        example: 
      - in: query
        name: sortOrder
        description: sortOrder
        schema:
          type: string
          enum:
          - Ascending
          - Descending
        required: false
      responses:
        '200':
          description: Reference data for mandate
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PagedMandate"
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
    post:
      tags:
      - Mandate
      summary: Creates a new mandate
      description: Creates a new mandate
      operationId: PostCollectionAsync
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CollectionCreationCommand"
      responses:
        '200':
          description: Entry was successfully created
          content:
            application/json:
              schema:
                description: Unique identifier of the created entry
                type: object
                required:
                - id
                properties:
                  id:
                    description: Unique identifier of the created entry
                    type: string
                    format: uuid
                    example: e4ec92f2-a375-4574-b6de-235a9fe13028
        '400':
          description: The query contains invalid arguments
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                InvalidRequestBody:
                  value:
                    errorType: InvalidRequestBody
                    logReference: '12345'
                    message: The code field is required.
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                CompanyForbidden:
                  value:
                    errorType: CompanyForbidden
                    logReference: '12345'
                    message: The 'erpId' must be part of your portfolio and have active
                      engagement
        '404':
          description: Elements of the query are not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                CompanyNotFound:
                  value:
                    errorType: CompanyNotFound
                    logReference: '12345'
                    message: The 'erpId' account number has not been found
                BankNotFound:
                  value:
                    errorType: BankNotFound
                    logReference: '12345'
                    message: The 'xxxx' bank code has not been found
  "/mandate/{mandateId}/unsigned":
    get:
      tags:
      - Mandate
      summary: Downloads an unsigned mandate
      description: Downloads an unsigned mandate
      operationId: DownloadUnsignedAsync
      parameters:
      - in: path
        name: mandateId
        description: The mandate identifier (UUID) given by /mandate
        schema:
          type: string
          format: uuid
        required: true
        example: e4ec92f2-a375-4574-b6de-235a9fe13028
      responses:
        '200':
          description: The unsigned mandate file (PDF format)
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                CompanyForbidden:
                  value:
                    errorType: CompanyForbidden
                    logReference: '12345'
                    message: The mandate belongs to a company that must be part of
                      your portfolio and have active engagement
        '404':
          description: Mandate not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                MandateNotFound:
                  value:
                    errorType: MandateNotFound
                    logReference: '12345'
                    message: The mandate with identifier 'mandateId' has not been
                      found
  "/mandate/{mandateId}/signed":
    post:
      tags:
      - Mandate
      summary: Uploads a signed mandate as form-data
      description: Uploads a signed mandate PDF file using multipart/form-data
      operationId: UploadSignedMandateAsync
      parameters:
      - in: path
        name: mandateId
        description: The mandate identifier (UUID) given by /mandate
        schema:
          type: string
          format: uuid
        required: true
        example: e4ec92f2-a375-4574-b6de-235a9fe13028
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The signed mandate PDF file
        required: true
      responses:
        '204':
          description: Upload was successful
        '400':
          description: Invalid mandate status
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                InvalidMandateStatus:
                  value:
                    errorType: InvalidMandateStatus
                    logReference: '12345'
                    message: The mandate status doesn't currently allow a file upload
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                CompanyForbidden:
                  value:
                    errorType: CompanyForbidden
                    logReference: '12345'
                    message: The mandate belongs to a company that must be part of
                      your portfolio and have active engagement
        '404':
          description: Mandate not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                MandateNotFound:
                  value:
                    errorType: MandateNotFound
                    logReference: '12345'
                    message: The mandate with identifier 'mandateId' has not been
                      found
  "/mandate/{mandateId}/deactivate":
    post:
      tags:
      - Mandate
      summary: Requests deactivation for the given mandate
      description: Requests deactivation for the given mandate
      operationId: DeactivateAsync
      parameters:
      - in: path
        name: mandateId
        description: The mandate identifier (UUID) given by /mandate
        schema:
          type: string
          format: uuid
        required: true
        example: e4ec92f2-a375-4574-b6de-235a9fe13028
      responses:
        '204':
          description: Deactivation was successful
        '400':
          description: Invalid mandate status
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                InvalidMandateStatus:
                  value:
                    errorType: InvalidMandateStatus
                    logReference: '12345'
                    message: The mandate status doesn't currently allow a deactivation
        '401':
          "$ref": "#/components/responses/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                CompanyForbidden:
                  value:
                    errorType: CompanyForbidden
                    logReference: '12345'
                    message: The mandate belongs to a company that must be part of
                      your portfolio and have active engagement
        '404':
          description: Mandate not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              examples:
                MandateNotFound:
                  value:
                    errorType: MandateNotFound
                    logReference: '12345'
                    message: The mandate with identifier 'mandateId' has not been
                      found
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Bearer token is missing or invalid
  schemas:
    Address:
      description: Data specific to the address
      type: object
      properties:
        street:
          description: Defines the signatory street address
          type: string
          example: RUE DU COMMERCE
        addressComplement:
          description: Defines the signatory address complements
          type: string
          example: 
        zipCode:
          description: Defines the signatory address zip code
          type: string
          example: '59590'
        city:
          description: Defines the signatory address city
          type: string
          example: RAISMES
        country:
          description: Defines the signatory address country
          type: string
          example: FRANCE
...
    Bban:
      description: French BBAN (RIB)
      type: object
      properties:
        bankCode:
          description: The bank code of the BBAN (first parameter)
          type: string
          required: true
          example: '12345'
        branchCode:
          description: The branch code of the BBAN (second parameter)
          type: string
          required: true
          example: '67890'
        accountNumber:
          description: The account number of the BBAN (third parameter)
          type: string
          required: true
          example: '12345678901'
        checkDigits:
          description: The check digits of the BBAN (last parameter)
          type: string
          required: true
          example: '23'
    CollectionCreationCommand:
      description: Data needed to create a new mandate
      type: object
      properties:
        erpId:
          description: Defines the identifier of the company
          type: string
          required: true
          example: '1000265308'
        signatory:
          "$ref": "#/components/schemas/Signatory"
        address:
          "$ref": "#/components/schemas/Address"
        bban:
          "$ref": "#/components/schemas/Bban"
    CollectionSummary:
      description: Summary of a data collection ; this object is intended to be used
        in tables
      type: object
      properties:
        id:
          description: Defines the mandate identifier
          type: string
          format: uuid
          example: e4ec92f2-a375-4574-b6de-235a9fe13028
        erpId:
          description: Defines the identifier of the company in the ERP
          type: string
          example: '1000265308'
        companyName:
          description: Defines the company legal name
          type: string
          example: Weyland Corporation
        accountNumber:
          description: Defines the bank account number of the mandate BBAN
          type: string
          example: '66842913356'
        bankName:
          description: Defines the name of the bank attached to the mandate
          type: string
          example: Crédit Agricole
        creationDate:
          description: Defines the creation date of the mandate
          type: string
          example: '2023-08-14T13:12:31.926Z'
        modificationDate:
          description: Defines the last update date of the mandate
          type: string
          example: '2023-08-14T13:12:31.926Z'
        statusCode:
          description: Defines the pulse status code of the collection mandate
          type: integer
          example: 40
...
    Error:
      type: object
      description: Represents an error when calling the Mandate API
      required:
      - errorType
      - logReference
      - message
      properties:
        errorType:
          description: Identifier of the error
          type: string
          example: InvalidCompanyCode
        logReference:
          description: Reference of the detailed log in Azure Insight
          type: string
          example: 12345abcd
        message:
          description: Error message. Can be displayed on client side
          type: string
          example: The specified company identifier is incorrect
...
    PagedCounters:
      description: Contains query metadata for a paged result
      type: object
      properties:
        all:
          description: Total number of matching mandates
          type: integer
          example: 32
        status10:
          description: Total number of mandates with status = 10 (Incident)
          type: integer
          example: 1
        status20:
          description: Total number of mandates with status = 20 (ToDo)
          type: integer
          example: 3
        status30:
          description: Total number of mandates with status = 30 (InProgress)
          type: integer
          example: 4
        status40:
          description: Total number of mandates with status = 40 (Active)
          type: integer
          example: 24
        status50:
          description: Total number of mandates with status = 50 (Inactive)
          type: integer
          example: 0
    PagedMandate:
      description: Contains one page of results + query metadata
      type: object
      properties:
        counters:
          "$ref": "#/components/schemas/PagedCounters"
        data:
          type: array
          items:
            "$ref": "#/components/schemas/CollectionSummary"
    Signatory:
      description: Data specific to the signatory
      type: object
      properties:
        title:
          description: Defines the signatory title
          type: string
          example: M
        firstName:
          description: Defines the signatory first name
          type: string
          example: maroo
        lastName:
          description: Defines the signatory Last name
          type: string
          example: elleuch
        email:
          description: Defines the signatory email
          type: string
          example: elleuch@email.com